% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_metrics.R
\name{ind_tree}
\alias{ind_tree}
\title{Tree variables to point cloud}
\usage{
ind_tree(las)
}
\arguments{
\item{las}{las object}
}
\description{
This function creates a point cloud with tree merics that can be summarized
}
\examples{
las <- readLAS(LASfile)
las_tree <- segment_trees(las, li2012())
las_tree <- ind_tree(las_tree)

}
\keyword{lidar}
\keyword{metrics}
\keyword{tree}
