% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_functions.R
\name{ind_vox}
\alias{ind_vox}
\title{Voxelize function}
\usage{
ind_vox(las, res = res)
}
\arguments{
\item{las}{las object}

\item{res}{voxel cubic resolution in units of las object}
}
\description{
This function provides individual voxel-based variables including sub-voxel density (SVi), frequency ratio (FRDi),
number of returns below each voxel (PDi), and the number of returns above each voxel (PDi_above).
}
\examples{
las <- readLAS(LASfile)
las_vox <- ind_vox(las, res = 2)
}
\keyword{lidar}
\keyword{metrics}
\keyword{voxel}
