% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_las.R
\docType{data}
\name{las}
\alias{las}
\title{Standard voxel metrics function}
\format{
An object of class \code{LAS} of length 1.
}
\usage{
las
}
\arguments{
\item{las}{an las file (made for small, 0.01 ha, areas)}

\item{resolution}{the size of voxels wanted for a majority of metrics unless defined by}

\item{sf_poly}{sf polygon to clip voxels to. Required.
different height bins in Pearse et al. 2019.}
}
\description{
This function provides a variety of voxel based metrics based off similar metrics in
Pearse et al. 2019 and Kim et al. 2016. The function also includes
additional summary statistics for height and intensity values within each voxel
}
\examples{
std_voxel()
}
\keyword{lidar}
\keyword{metrics}
\keyword{voxel}
