% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blackburn_et_al_2021_functions.R
\name{std_trees}
\alias{std_trees}
\title{Standard tree metrics function}
\usage{
std_trees(trees)
}
\arguments{
\item{trees}{A 'SpatialPointsDataFrame' created from lidR tree_metrics()}
}
\description{
This function takes the output from tree_metrics() in lidR and creates area based summaries from the segmented trees.
}
\keyword{lidar}
\keyword{metrics}
\keyword{tree}
