% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_las.R
\name{std_voxel_all}
\alias{std_voxel_all}
\title{Voxel metrics for individual voxels}
\usage{
std_voxel_all(
  las,
  resolution,
  vox_ht = max(vox$Z),
  vox_x = c(min(vox$X), max(vox$X)),
  vox_y = c(min(vox$Y), max(vox$Y))
)
}
\arguments{
\item{las}{an las file (made for small, 0.01 ha, areas).}

\item{resolution}{the square dimensions of voxels.}

\item{vox_ht}{max height of voxels on plot. This is best to be used as the max height of the trees wihtin plots}

\item{vox_x}{is a combination of min and max x of plot. Default is  c(min(vox$X), max(vox$X))}

\item{vox_y}{is a combination of min and max x of plot. Default is  c(min(vox$Y), max(vox$Y))}
}
\description{
This function provides a variety of voxel based metrics based off similar metrics in
Pearse et al. 2019 and Kim et al. 2016. The function also includes
additional summary statistics for height and intensity values within each voxel
}
\examples{
std_voxel()
}
\keyword{lidar}
\keyword{metrics}
\keyword{voxel}
