% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidRattR_functions.R
\name{tree_raster}
\alias{tree_raster}
\title{Tree raster summary function}
\usage{
tree_raster(las, rast_res)
}
\arguments{
\item{las}{las object}
}
\description{
This function provides individual tree-based variables including number of trees (ntrees),
height of trees (ht), number of points per tree (npts), and crown area (ca). These variables are summarized
within each pixel including the mean, median, variance, standard deviation, coefficient of variation,
IQR, skewness, and kurtosis
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile)
las <- filter_poi(las, Classification != LASNOISE)
las <- normalize_height(las, knnidw(k = 8, p = 2))
las <- filter_poi(las, Z < 50 & Z >= 1.37 )
las <- decimate_points(las, random_per_voxel(res = 1, n = 8))
las_tree <- segment_trees(las, li2012())
tree_metrics <- tree_raster(las_tree)
plot(tree_metrics)
}
\keyword{lidar}
\keyword{metrics}
\keyword{tree}
