% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidRattR_functions.R
\name{tree_sum}
\alias{tree_sum}
\title{Tree-based variable summary statistics}
\usage{
tree_sum(X, Y, Z, npoints, ca)
}
\arguments{
\item{columns}{from the lidR::crown_metrics() function (i.e., X, Y, Z, npoints, ca)}
}
\description{
This function provides summarized tree-based variables including the mean, standard deviation, and coefficient of variation.
Summaries for this function are done to the entire area provided either directly or within the
lidR::pixel_metrics() function.
}
\keyword{lidar}
\keyword{tree}
