% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_functions.R
\name{tree_sum_raster}
\alias{tree_sum_raster}
\title{Tree-based variable summary statistics for lidR::pixel_metrics() function}
\usage{
tree_sum_raster(Z, npoints, ca)
}
\description{
This function provides summarized tree-based variables just as the tree_sum() function.
This can be used in conjunction is lidR::pixel_metrics() to make raster layers of tree-based summary statistics.
}
\examples{
las <- readLAS(LASfile)
las_tree <- segment_trees(las, li2012())
las_tree <- ind_tree(las_tree)
t_metrics <- pixel_metrics(las_tree, tree_sum_raster(Z, npoints, ca), res = 20)

}
\keyword{lidar}
\keyword{metrics}
\keyword{tree}
