% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_las.R, R/metrics_las_working.R
\name{vox_mt}
\alias{vox_mt}
\title{Voxel metrics function}
\usage{
vox_mt(z, i)

vox_mt(z, i)
}
\arguments{
\item{z}{height}

\item{i}{intensity}
}
\description{
This function takes the intensity and height values for each lidar point and
is used within lidR::voxel_metrics() which is used within lidRmts::std_voxel()

This function takes the intensity and height values for each lidar point and
is used within lidR::voxel_metrics() which is used within lidRmts::std_voxel()
}
\examples{
voxels <- voxel_metrics(las, func = vox_mt(Z, as.numeric(Intensity)), res = resolution)
voxels <- voxel_metrics(las, func = vox_mt(Z, as.numeric(Intensity)), res = resolution)
}
\keyword{lidar}
\keyword{metrics}
\keyword{voxel}
