% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_functions.R
\name{vox_mt2}
\alias{vox_mt2}
\title{Individual voxel-based variable function}
\usage{
vox_mt2(vox, res)
}
\arguments{
\item{vox}{lasmetrics3d object created by lidR::voxel_metrics() using the vox_mt() function}
}
\description{
This function provides individual voxel-based variables including the frequency ratio (FRDi),
number of returns below each voxel (PDi), and the number of returns above each voxel (PDi_above),
}
\examples{
std_voxel()

}
\keyword{lidar}
\keyword{metrics}
\keyword{voxel}
