% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_metrics.R
\name{vox_sum}
\alias{vox_sum}
\title{Voxel-based variable summary statistics}
\usage{
vox_sum(las_vox, res)
}
\arguments{
\item{res}{voxel cubic resolution in units of las object}

\item{las}{voxelized las object}
}
\description{
This function provides summarized voxel-based variables including for each of the variables created using the
vox() function. Summary statistics include the mean, median, varaince, standard deviation, coefficient of variation,
IQR, skewness, and kurtosis.
}
\examples{
las <- readLAS(LASfile)
las_vox <- vox(las, res = 2)
voxel_summaries <- vox_sum(las_vox, 2)

}
\keyword{lidar}
\keyword{metrics}
\keyword{voxel}
