% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_metrics.R
\name{vox_sum_raster}
\alias{vox_sum_raster}
\title{Voxel-based variable summary statistics for lidR::pixel_metrics() function}
\usage{
vox_sum_raster(SVi, FRDi, PDi, PDi_above)
}
\description{
This function provides summarized voxel-based variables just as the vox_sum() function.
This can be used in conjunction is lidR::pixel_metrics() to make raster layers of voxel-based summary statistics.
}
\examples{
las <- readLAS(LASfile)
las_vox <- vox(las, res = 2)
v_metrics <- pixel_metrics(las_vox, func = .vox_raster, res = 20)


}
\keyword{lidar}
\keyword{metrics}
\keyword{voxel}
