% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidRattR_functions.R
\name{voxel_summary}
\alias{voxel_summary}
\title{Voxel point cloud summary function
This function provides individual voxel-based variables including the frequency ratio (FRDi),frequency ratio (FRSVi),
number of returns below each sub-voxel (SVi) (PDi_below), and the number of returns above each SVi (PDi_above),
sum of returns within a sub-voxels (SV_sum), classifcation of a sub-voxel maximum (SVM), and
median height of the number of returns above (SVM_md_above) or within (SVM_md) SVMs. These variables are summarized
for the entir point cloud including the mean, median, variance, standard deviation, coefficient of variation,
IQR, skewness, and kurtosis}
\usage{
voxel_summary(las, vox_res = vox_res, max_z)
}
\arguments{
\item{las}{las object}

\item{vox_res}{SVi cubic resolution in units of las object}

\item{max_z}{If numeric, the final height for all sub-voxels. If NULL, the max height for the entire point cloud
(or chunk for catalog processing) will be used.}
}
\description{
Voxel point cloud summary function
This function provides individual voxel-based variables including the frequency ratio (FRDi),frequency ratio (FRSVi),
number of returns below each sub-voxel (SVi) (PDi_below), and the number of returns above each SVi (PDi_above),
sum of returns within a sub-voxels (SV_sum), classifcation of a sub-voxel maximum (SVM), and
median height of the number of returns above (SVM_md_above) or within (SVM_md) SVMs. These variables are summarized
for the entir point cloud including the mean, median, variance, standard deviation, coefficient of variation,
IQR, skewness, and kurtosis
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las <- readLAS(LASfile)
las = filter_poi(las, Classification != LASNOISE)
las = normalize_height(las, knnidw(k = 8, p = 2))
las = filter_poi(las, Z < 50 & Z >= 1.37 )
las <- decimate_points(las, random_per_voxel(res = 1, n = 8))
metrics <- voxel_summary(las, vox_res = 2, max_z = 40)
print(metrics)
}
\keyword{lidar}
\keyword{metrics}
\keyword{voxel}
