% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blackburn_et_al_2021_functions.R
\name{std_cloud}
\alias{std_cloud}
\title{Standard cloud metrics function}
\usage{
std_cloud(x, y, z, i, ReturnNumber, usei)
}
\arguments{
\item{x}{X location.}

\item{y}{Y location.}

\item{z}{Z location.}

\item{ReturnNumber}{Return number column from las object.}
}
\description{
This function provides a variety of area based metrics from Pearse et al. 2019 and Mitchell et al. 2012.
}
\examples{
std_cloud()
}
\keyword{cloud}
\keyword{lidar}
\keyword{metrics}
