% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_las.R
\name{std_trees}
\alias{std_trees}
\title{Standard tree metrics function}
\usage{
std_trees(trees)
}
\arguments{
\item{trees}{A 'SpatialPointsDataFrame' created from lidR tree_metrics()}
}
\description{
This function takes the output from tree_metrics() in lidR and creates area based summaries from the segmented trees.
}
\examples{
trees <- lastrees(las.t, watershed(chm))
trees <- tree_metrics(trees, func = .stdtreemetrics )
tree_stats <- std_trees(trees)
}
\keyword{lidar}
\keyword{metrics}
\keyword{tree}
