% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidRattR_functions.R
\name{tree_sum}
\alias{tree_sum}
\title{Tree-based variable summary statistics}
\usage{
tree_sum(las_tree)
}
\arguments{
\item{las}{las object from ind_tree() function}
}
\description{
This function provides summarized tree-based variables  for each of the variables created using the
ind_tree() function. Summary statistics include the mean, standard deviation, and coefficient of variation.
}
\examples{
las <- readLAS(LASfile)
las_tree <- segment_trees(las, li2012())
las_tree <- ind_tree(las_tree)
plot_t_metrics <- tree_sum(las_tree)
}
\keyword{lidar}
\keyword{metrics}
\keyword{tree}
