% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidRattR_functions.R
\name{vox_mt}
\alias{vox_mt}
\title{Individual voxel-based SVi function}
\usage{
vox_mt(z)
}
\arguments{
\item{z}{height}
}
\description{
This function takes the height values (Z) for each lidar point point cloud and
is used within lidR::voxel_metrics() to calculate the sub-voxel i (SVi) density (SViD) and a list of Z values for for each sub-voxel.
}
\examples{
voxels <- voxel_metrics(las, func = vox_mt(Z), res = 2)
}
\keyword{lidar}
\keyword{voxel}
