% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blackburn_et_al_2021_functions.R,
%   R/updated_metrics.R
\name{vox_mt}
\alias{vox_mt}
\title{Voxel metrics function}
\usage{
vox_mt(z)

vox_mt(z)
}
\arguments{
\item{z}{height}

\item{i}{intensity}
}
\description{
This function takes the intensity and height values for each lidar point and
is used within lidR::voxel_metrics() which is used within lidRmts::std_voxel()

This function takes the height values for each lidar point point cloud and
is used within lidR::voxel_metrics() to calculate the sub-voxel density (SVi) for each voxel.
}
\examples{
voxels <- voxel_metrics(las, func = vox_mt(Z, as.numeric(Intensity)), res = resolution)
voxels <- voxel_metrics(las, func = vox_mt(Z), res = 2)
}
\keyword{lidar}
\keyword{metrics}
\keyword{voxel}
