% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidRattR_functions.R
\name{vox_mt2}
\alias{vox_mt2}
\title{Individual voxel-based variable function}
\usage{
vox_mt2(vox, res, max_z)
}
\arguments{
\item{vox}{lasmetrics3d object created by lidR::voxel_metrics() using the vox_mt() function}

\item{res}{SVi cubic resolution in units of las object}

\item{max_z}{if numeric, the final height for all sub-voxels}
}
\description{
This function provides individual voxel-based variables including the frequency ratio (FRDi),frequency ratio (FRSVi),
number of returns below each sub-voxel (SVi) (PDiBelow), and the number of returns above each SVi (PDiAbove),
sum of returns within a sub-voxels (SVsum), classifcation of a sub-voxel maximum (SVM),
median height of the number of returns above (SVMmdAbove) or within (SVMmd) SVMs
}
\keyword{lidar}
\keyword{voxel}
