% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidRattR_functions.R
\name{vox_sum}
\alias{vox_sum}
\title{Voxel-based variable summary statistics}
\usage{
vox_sum(
  SVi_D,
  FRDi,
  FRSVi,
  PDi_below,
  PDi_above,
  SV_sum,
  SVM_md_above,
  SVM_md,
  SVM,
  vox_res
)
}
\description{
This function provides summarized voxel-based variables including the mean, median, variance, standard deviation, coefficient of variation,
IQR, skewness, and kurtosis. Summaries for this function are done to the entire area provided either directly or within the
lidR::pixel_metrics() function.
}
\keyword{lidar}
\keyword{metrics}
\keyword{voxel}
